requirements.txt
pydiscourse>=1.3,<1.4
PyYAML>=6.0,<6.1
